+++ src/main/java/org/openbase/bco/bcozy/permissions/model/RecursiveUnitConfig.java
@@ -1,5 +1,6 @@
 package org.openbase.bco.bcozy.permissions.model;
 
+import com.jfoenix.controls.JFXTreeTableView;
 import com.jfoenix.controls.datamodels.treetable.RecursiveTreeObject;
 import javafx.beans.property.SimpleObjectProperty;
 import rst.domotic.unit.UnitConfigType;
@@ -7,6 +8,8 @@ import rst.domotic.unit.UnitConfigType;
 import java.util.Objects;
 
 /**
+ * Wrapper around {@link UnitConfigType.UnitConfig}, used for displaying in {@link JFXTreeTableView}
+ *
  * @author vdasilva
  */
 public class RecursiveUnitConfig extends RecursiveTreeObject<RecursiveUnitConfig> {
