+++ src/main/java/org/openbase/bco/bcozy/permissions/model/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Contains the model used to display and edit unit-permissions.
+ * <p>
+ * Mapping between bco-Permissions ans this model happens in {@link org.openbase.bco.bcozy.permissions.model.PermissionsServiceImpl}.
+ *
+ * @author vdasilva
+ */
+package org.openbase.bco.bcozy.permissions.model;
